requirements.txt
streamlit
python-pptx
pandas
openpyxl
sentence-transformers
scikit-learn
pyyaml
torch
